requirements.txt

# Core ADK package
google-adk>=1.0.0

# Google Generative AI (for Gemini models)
google-generativeai>=0.3.0

# Python environment management (for loading .env files)
python-dotenv>=1.0.0

# Optional: For production deployment
# uvicorn>=0.20.0  # For FastAPI server
# fastapi>=0.100.0  # For API endpoints
